+++ probemanager/api/urls.py
@@ -12,7 +12,6 @@ router = routers.DefaultRouter()
 
 router.register(r'^admin/users', views.UserViewSet)
 router.register(r'^admin/groups', views.GroupViewSet)
-router.register(r'^rules/classtype', views.ClassTypeViewSet)
 router.register(r'^core/server', views.ServerViewSet, base_name="core")
 router.register(r'^core/sshkey', SshKeyView, base_name="core")
 router.register(r'^core/configuration', views.ConfigurationViewSet, base_name="core")