+++ CHANGELOG.rst
@@ -1,12 +1,30 @@
+=========
 Changelog
 =========
 
 All notable changes to this project will be documented in this file.
 
-The format is based on `Keep a Changelog`_ and this project adheres to `Semantic Versioning`_.
-
 [Unreleased]
 ------------
+
+[1.2.0] - 2018-04-30
+--------------------
+
+Added
+~~~~~
+
+-  Add Classtype to Suricata.
+
+Changed
+~~~~~~~
+
+- Improved API.
+- Improved Documentation.
+
+Removed
+~~~~~~~
+
+-  Remove Classtype from rules
 
 
 [1.1.0] - 2018-04-19